digitalfacets_requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-admin-sortable2==0.7.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
Markdown==3.1
Pillow==6.0.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
